requirements.txt
amqp==2.2.2
asgi-redis==1.4.2
asgiref==1.1.2
astroid==1.5.3
attrs==17.2.0
autobahn==17.9.1
Automat==0.6.0
backports.functools-lru-cache==1.4
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
channels==1.1.6
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
constantly==15.1.0
coverage==4.4.1
daphne==1.3.0
dj-database-url==0.4.2
Django==2.0.1
django-braces==1.12.0
django-countries==4.6.1
django-filter==1.0.4
django-rest-framework==0.1.0
djangorestframework==3.7.7
enum34==1.1.6
gunicorn==19.7.1
haikunator==2.1.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
isort==4.2.15
kombu==4.1.0
lazy-object-proxy==1.3.1
Markdown==2.6.9
mccabe==0.6.1
msgpack-python==0.4.8
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.1
pylint==1.7.2
pytz==2017.2
redis==2.10.6
requests==2.18.4
singledispatch==3.4.0.3
six==1.10.0
Twisted==17.5.0
txaio==2.8.2
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1
wrapt==1.10.11
zope.interface==4.4.2
